Several changes:
 - WLineEdit: That should also be std::u32string instead of std::wstring (for Windows)
 - .? should just become ?, count components in fixRelativeUrl instead of slashes
 - Removed sourceforge and mailing list stuff from wt homepage
 - Optimized WTreeView::modelDataChanged() to use nodeForIndex()
 - Don't left shift too far, cast to int64_t first
 - Direct2D WRasterImage implementation + small fixes
 - Fixed getPixel/getPixels in WRasterImage-gm with non-8-bit quantum

+++ src/http/RequestParser.C
@@ -522,7 +522,7 @@ RequestParser::parseWebSocketMessage(Request& req, ReplyPtr reply,
 
       break;
     case ws00_binary_length:
-      if (remainder_ >= (0x01 << 56)) {
+      if (remainder_ >= (((int64_t)0x01) << 56)) {
         LOG_ERROR("ws: oversized binary frame: overflows 64-bit signed integer");
         return Request::Error;
       }
